Move identification helper functions.
